% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bw_gr.R
\name{create_bw_gr}
\alias{create_bw_gr}
\title{Create coverage GRanges from bigwig file with specific genome ranges}
\usage{
create_bw_gr(plot_region, bw_path, se_spec_table)
}
\arguments{
\item{plot_region}{a string of region of interest}

\item{bw_path}{a vector of bigwig file path}

\item{se_spec_table}{se specifity table generated from "search_db" function}
}
\value{
A list of bigwig Granges for each bigwig file and plot region gr
}
\description{
A function to create a coverage Granges of bigwig file
}
\details{
This function will return a list of coverage granges of multiple bigwig file at specific regions
}
\examples{
# bigwig file path
bw_file_path <- c("merged_normalized_bw/786-0_normalized.bw","merged_normalized_bw/A549_normalized.bw")
plot_region <- "chr1_1109053_1176954"

bw_gr <- create_bw_gr(plot_region,bw_file_path)
}
