% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gviz_tracks.R
\name{create_gviz_tracks}
\alias{create_gviz_tracks}
\title{Create Gviz plot tracks for specific regions}
\usage{
create_gviz_tracks(
  bw_gr_list,
  se_spec_table,
  cell = "specific",
  plot_region,
  txdb = "none"
)
}
\arguments{
\item{bw_gr_list}{list of bigwig Granges generated from "create_be_gr" function}

\item{se_spec_table}{se specifity table generated from "search_db" function}

\item{cell}{a comman separated string of which cell to plot.
Examples:"cell1,cell2,cell3","specific","specific,cell1,cell2".
(Default: specific. Cell have specific CEs within the plot region.)}

\item{plot_region}{a string indicate region of interest. (Example:"chr1_1109053_1176954"or "chr1:112388326-112392055")}

\item{txdb}{a txdb file for gene annotation track. (Default: "none" No gene annotation track)}
}
\value{
A output list with different type of gviz tracks:
htrack: major data tracks with highlighted specific CEs.
atrack_ce: annotation track indicating CE locations.
atrack_se: annotation track indicating SE locations overlap with plot region.
atrack_query: annotation track indicating plot region location.
gtrack: gene annotation track.
plot_region: vector of plot region information (chr,start,end).
axistrack: genome axis track.
}
\description{
A function to create Gviz data tracks and annotation tracks with specific CEs
}
\details{
This function will create different Gviz data tracks based on the selected cell and
annotation tracks of specific CEs and SEs
}
\examples{
# region of interests
plot_region <- "chr1_1109053_1176954"

# search db to find overlap specific SEs
search_db_table <- search_db(plot_region,query_type="se_region",cSEAdb)

# create bw_gr list
bw_gr <- create_bw_gr(plot_region,bigWigs$V1)

# default: only plot specific cells with no gene track
tracks_default <- create_gviz_tracks(bw_gr_list=bw_gr,
                             se_spec_table=search_db_table,
                             plot_region=plot_region)

# load txdb file
library(TxDb.Hsapiens.UCSC.hg38.knownGene)
# plot selected cells within bw_gr and gene track from txdb object
tracks <- create_gviz_tracks(bw_gr_list=bw_gr,
                             se_spec_table=search_db_table,
                             cell="MCF7,A549,K-562,specific",
                             plot_region=plot_region,
                             txdb=TxDb.Hsapiens.UCSC.hg38.knownGene)

# plot tracks using Gviz
plotTracks(c(plot_out$htrack,plot_out$atrack_ce,
             plot_out$atrack_se,plot_out$gtrack,plot_out$atrack_query,
             plot_out$axistrack),
           chr=plot_out$plot_region[1],
           from=as.integer(plot_out$plot_region[2]),
           to=as.integer(plot_out$plot_region[3]))
}
