% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_db.R
\name{search_db}
\alias{search_db}
\title{Search database with query and return SE specificity}
\usage{
search_db(query, query_type = "se_region", se_db, gene_region = F)
}
\arguments{
\item{query}{a query to search.
Can be SE regions (with format of "chr_start_end"), SE bed file, cell name,
cancer name, or tissue name.(The full list of cell name, cancer name and tissue name
available in this database can be found in the description file)}

\item{query_type}{a character parameter indicates which type of query it is:
"se_region", "se_bed", "gene", "cancer", or "cell". (Default: se_region)}

\item{se_db}{the cSEAdb object loaded from the package}
}
\value{
A list of SE regions and there cell and cancer specificity will return.
}
\description{
A function to return the overlapped SE regions with a SE query string or SE bed file
}
\details{
This function will return the SE regions in the database which are overlapped
with a query SE region.
}
\examples{

# search with SE string
hit_se <- search_db(c("chr1_121000_123250","chr5_111111_222222"),query_type="se_region")

# search with SE bed file
hit_se <- search_db("se_query_list.bed",query_type="se_bed")

# search with gene symbol
hit_se <- search_db(c("MYC","FOXA2"),query_type="gene")

# search with cell name
hit_se <- search_db(c("HL-60","MCF7"),query_type="cell")

# search with cancer name
hit_se <- search_db(c("Amelanotic melanoma","Prostate carcinoma"),query_type="cancer")

}
